Fix issue with carousel
Modified the function that scrolls. When using carousel, buttons PREV / NEXT did not work, because the href that was used to slide to the next image (e.g. "<a href="#myCarousel>") was hijacked by the scroll function. Added exception for [data-slide] attribute and a few others.

For more info, see link:
https://www.drupal.org/node/2233127